pyproject.toml
[tool.poetry]
name = "paper-versus-people"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "paper_versus_people"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-model-utils = "^4.3.1"
psycopg2 = "^2.9.5"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
